Moved mail sending to utility class
