+++ src/main/java/dev/dragoncave/yap/backend/rest/security/MailSend.java
@@ -1,0 +1,45 @@
+package dev.dragoncave.yap.backend.rest.security;
+
+import javax.mail.*;
+import javax.mail.internet.InternetAddress;
+import javax.mail.internet.MimeBodyPart;
+import javax.mail.internet.MimeMessage;
+import javax.mail.internet.MimeMultipart;
+import java.util.Properties;
+
+public class MailSend {
+	public static void sendMail(String from, String recipient, String subject, String messageContent) throws MessagingException {
+		Properties prop = new Properties();
+		prop.put("mail.smtp.auth", true);
+		prop.put("mail.smtp.starttls.enable", "true");
+		prop.put("mail.smtp.host", "smtp.gmail.com");
+		prop.put("mail.smtp.port", "587");
+		prop.put("mail.smtp.ssl.trust", "smtp.gmail.com");
+
+		String username = "yapreset@gmail.com";
+		String password = System.getenv("MAIL_PASS");
+
+		Session session = Session.getInstance(prop, new Authenticator() {
+			@Override
+			protected PasswordAuthentication getPasswordAuthentication() {
+				return new PasswordAuthentication(username, password);
+			}
+		});
+
+		Message message = new MimeMessage(session);
+		message.setFrom(new InternetAddress(from));
+		message.setRecipients(
+				Message.RecipientType.TO, InternetAddress.parse(recipient));
+		message.setSubject(subject);
+
+		MimeBodyPart mimeBodyPart = new MimeBodyPart();
+		mimeBodyPart.setContent(messageContent, "text/html");
+
+		Multipart multipart = new MimeMultipart();
+		multipart.addBodyPart(mimeBodyPart);
+
+		message.setContent(multipart);
+
+		Transport.send(message);
+	}
+}
