+++ src/main/java/dev/dragoncave/yap/backend/rest/security/controllers/SecurityController.java
@@ -3,6 +3,7 @@ package dev.dragoncave.yap.backend.rest.security.controllers;
 import dev.dragoncave.yap.backend.databasemanagers.PasswordController;
 import dev.dragoncave.yap.backend.databasemanagers.UserController;
 import dev.dragoncave.yap.backend.rest.objects.User;
+import dev.dragoncave.yap.backend.rest.security.MailSend;
 import dev.dragoncave.yap.backend.rest.security.PasswordUtils;
 import dev.dragoncave.yap.backend.rest.security.tokens.DatabaseTokenStore;
 import dev.dragoncave.yap.backend.rest.security.tokens.TokenUtils;
@@ -64,40 +65,12 @@ public class SecurityController {
 			String resetCode = TokenUtils.generateToken(48);
 			PasswordController.insertPasswordResetCode(requestBody.get("emailAddress"), resetCode);
 
-			Properties prop = new Properties();
-			prop.put("mail.smtp.auth", true);
-			prop.put("mail.smtp.starttls.enable", "true");
-			prop.put("mail.smtp.host", "smtp.gmail.com");
-			prop.put("mail.smtp.port", "587");
-			prop.put("mail.smtp.ssl.trust", "smtp.gmail.com");
-
-			String username = "yapreset@gmail.com";
-			String password = System.getenv("MAIL_PASS");
-
-			Session session = Session.getInstance(prop, new Authenticator() {
-				@Override
-				protected PasswordAuthentication getPasswordAuthentication() {
-					return new PasswordAuthentication(username, password);
-				}
-			});
-
-			Message message = new MimeMessage(session);
-			message.setFrom(new InternetAddress("yapreset@gmail.com"));
-			message.setRecipients(
-					Message.RecipientType.TO, InternetAddress.parse(requestBody.get("emailAddress")));
-			message.setSubject("YAP Password Reset");
-
-			String messageContent = "Your Reset code is: \n" + resetCode;
-
-			MimeBodyPart mimeBodyPart = new MimeBodyPart();
-			mimeBodyPart.setContent(messageContent, "text/html");
-
-			Multipart multipart = new MimeMultipart();
-			multipart.addBodyPart(mimeBodyPart);
-
-			message.setContent(multipart);
-
-			Transport.send(message);
+			MailSend.sendMail(
+					"yapreset@gmail.com",
+					requestBody.get("emailAddress"),
+					"YAP Password reset code",
+					"Your Reset code is: \n" + resetCode
+			);
 			return new ResponseEntity<>(HttpStatus.OK);
 		} catch (Exception e) {
 			e.printStackTrace();
